requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
python-multipart==0.0.*
python-dotenv==1.1.*
PyMuPDF==1.25.*
fpdf2==2.8.*
websockets==15.*
httpx>=0.27
transformers>=4.44
torch>=2.1
sentencepiece>=0.2
